Swap back on exception
If there is an exception in the middle of a swap block, swapper should
swap back to the original and continue to raise the error. The old
behavior would stay in the new database. If you swapped within a job in
resque, this would cause job failures to get posted to a different redis
server than from where they originated.